x/grid/types/message_submit_estimate_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"grid/testutil/sample"
)

func TestMsgSubmitEstimate_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSubmitEstimate
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSubmitEstimate{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSubmitEstimate{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
